view Order and OrderDetail

+++ LarsProjekt/Models/OrderDetailModel.cs
@@ -3,8 +3,8 @@
 public class OrderDetailModel
 {
     public long Id { get; set; }
-    public int OrderId { get; set; }
-    public int ProductId { get; set; }
+    public long OrderId { get; set; }
+    public long ProductId { get; set; }
     public int Quantity { get; set; }
     public decimal? UnitPrice { get; set; }
     public virtual ProductModel Product { get; set; }
